% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStocks.R
\name{calculateStocks}
\alias{calculateStocks}
\title{Calculate total stock size for a set of user-specified months, molecules, classes and patient stocks}
\usage{
calculateStocks(
  data,
  months = NULL,
  molecules = NULL,
  classes = NULL,
  stocks = NULL,
  ever_tried = TRUE
)
}
\arguments{
\item{data}{Dataframe to use containing patient_id, weight, month, molecule_name, class_name, and stock_name colunms}

\item{months}{Month to filter on, either exact match [i.e., 60] or bounded list [i.e., (24, 36) ]}

\item{molecules}{Set of molecules names to filter on}

\item{classes}{Set of drug classes names to filter on}

\item{stocks}{Set of patient stocks names to filter on}

\item{ever_tried}{Whether to count each patient once ("ever tried") or multiple times ("patient-months")}
}
\value{
Total sum of weights for the use-specified months, drugs, classes and/or stocks
}
\description{
Calculate total stock size for a set of user-specified months, molecules, classes and patient stocks
}
\examples{
 calculateStocks(zavus24_stocks_table, months=c(1,60), classes = c("CGRP Oral", "CGRP Injectable") )
}
