% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zavus24_stocks_table}
\alias{zavus24_stocks_table}
\title{Example of patient drugs/classes/stocks month-over-month for the Zavegepant US 2024 project}
\format{
A data frame with 43347 rows and 6 variables:
\describe{
  \item{patient_id}{patient ID}
  \item{weight}{patient projected}
  \item{month}{month}
  \item{molecule_name}{molecule the patient is on}
  \item{class_name}{drug class to which the molecule belongs to}
  \item{stock_name}{patient stock, as a function of the patient's highest molecules/classes on any given month}
}
}
\source{
{Created inhouse to be a used as an example.}
}
\usage{
zavus24_stocks_table
}
\description{
A dataset containing drugs for migraine patients in the US month over month over the oast 5 years.
}
\examples{
data(zavus24_stocks_table)  # Check Lazy Loading. It may only become visible once first called.
}
\keyword{datasets}
